make size float

+++ internal/godojs/godojs.go
@@ -808,7 +808,7 @@ func VolumeSnapshotToVM(vm *otto.Otto, g *godo.Snapshot) otto.Value {
 		"volume_id": g.ResourceID,
 		"name":      g.Name,
 		"regions":   g.Regions,
-		"size":      int64(g.SizeGigaBytes),
+		"size":      float64(g.SizeGigaBytes),
 	})
 }
 
@@ -822,7 +822,7 @@ func SnapshotToVM(vm *otto.Otto, g *godo.Snapshot) otto.Value {
 		"resource_type": g.ResourceType,
 		"name":          g.Name,
 		"regions":       g.Regions,
-		"size":          int64(g.SizeGigaBytes),
+		"size":          float64(g.SizeGigaBytes),
 		"min_disk_size": int64(g.MinDiskSize),
 		"created_at":    g.Created,
 	})
